+++ src/main/java/com/easysales/Repositories/CompraRepository.java
@@ -9,5 +9,5 @@ import java.util.UUID;
 @EnableJpaRepositories
 
 @Repository
-public interface CompraRepository extends JpaRepository<Compra, Integer> {
+public interface CompraRepository extends JpaRepository<Compra, Long> {
 }
