+++ src/main/java/com/easysales/Repositories/ProdutoRepository.java
@@ -8,4 +8,4 @@ import org.springframework.stereotype.Repository;
 import java.util.UUID;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto, Integer> {}
+public interface ProdutoRepository extends JpaRepository<Produto, Long> {}
